refactor(1.1): extraer constants, ui y banner a jobhunter/
- jobhunter/constants.py con BASE_DIR (ajustado al ubicar paquete),
  CONFIG_PATH, SESSION_DIR, KB_PATH, VERSION, GEMINI_MODELS,
  TIME_FILTERS, BACK, BANNER_LARGE, BANNER_SMALL
- jobhunter/ui.py con el singleton console de Rich
- jobhunter/banner.py con get_banner()
- job.py reemplaza definiciones por imports (sigue funcional)
- tests/test_banner.py cubre banner y constants
