+++ jobhunter/ui.py
@@ -1,0 +1,4 @@
+"""Console singleton de Rich. Todos los modulos importan `console` desde aqui."""
+from rich.console import Console
+
+console = Console()
